requirements.txt
# ======================== CORE ==========================
fastapi==0.115.13
uvicorn==0.34.3
starlette==0.46.2
python-dotenv==1.0.1

# ======================== ASYNC & CONCURRENCY ============
uvloop==0.21.0
anyio==4.9.0
sniffio==1.3.1
httpx==0.28.1
asyncpg==0.30.0
aiomysql==0.2.0
asyncmy==0.2.10
aiosmtplib==4.0.1

# ======================== DATABASE =======================
SQLAlchemy==2.0.41
PyMySQL==1.1.1
psycopg2-binary==2.9.10
Alembic==1.16.2

# ======================== AWS & STORAGE ===================
aioboto3==15.0.0

# ======================== BACKGROUND TASKS ===============
redis>=4.2.0,<6
arq==0.26.3

# ======================== SCHEDULING ======================
# apscheduler==3.11.0
apscheduler[sqlalchemy]==3.11.0

# ======================== STREAM PROCESSING (KAFKA) =======
# confluent-kafka==2.10.1
# confluent-kafka==2.3.0

# ======================== UTILITIES & HELPERS ============
python-dateutil==2.9.0.post0
pytz==2025.2
cachetools==6.1.0
python-slugify==8.0.4
text-unidecode==1.3
arrow==1.3.0
charset-normalizer==3.4.2
idna==3.10
urllib3==2.5.0
certifi==2025.6.15
six==1.17.0
click==8.2.1
asyncclick==8.1.8
orjson==3.10.18
python-multipart==0.0.7

# ======================== TEMPLATE ENGINE =================
jinja2==3.1.6
MarkupSafe==3.0.2

# ======================== SECURITY / JWT ==================
PyJWT==2.10.1

# ======================== LOGGING & OUTPUT ================
rich==13.9.4
rich-toolkit==0.13.2
Pygments==2.19.1
colorama==0.4.6

# ======================== VALIDATION ======================
pydantic[email]==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
annotated-types==0.7.0
typing_extensions==4.14.0
types-python-dateutil==2.9.0.20250516

# ======================== CLI / DEV TOOLS ================
typer==0.16.0
fastapi-cli==0.0.7
watchfiles==1.1.0
shellingham==1.5.4
dictdiffer==0.9.0
CacheControl==0.14.3
binaryornot==0.4.4
chardet==5.2.0
prometheus-fastapi-instrumentator==7.1.0

# ======================== HTTP & NETWORK ==================
h11==0.16.0
httptools==0.6.4
# ======================== MARKDOWN SUPPORT (DOCS) =========
markdown-it-py==3.0.0
mdurl==0.1.2


setuptools>=68.0.0
user-agents